% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffcar-package.R
\docType{package}
\name{ffcar-package}
\alias{ffcar-package}
\title{Functions for Creating Analysis Reports}
\description{
The \pkg{ffcar} package provides assorted
functions to help with a variety of common
tasks when creating analysis reports. See
\url{https://rettopnivek.github.io/ffcar/}
for more details.
}
