% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Reporting.R
\name{ffcar_add_code}
\alias{ffcar_add_code}
\title{Create RMarkdown Code for R Code}
\usage{
ffcar_add_code(chr_code, chr_tag = "", chr_options = ", echo=FALSE")
}
\arguments{
\item{chr_code}{A character vector with the lines of
R code to add.}

\item{chr_tag}{A character string, an optional code
segment label.}

\item{chr_options}{A character string, optional
R code segment options.}
}
\value{
A character vector.
}
\description{
Function that creates the code to add an R code
segment to a RMarkdown file.
}
