% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Reporting.R
\name{ffcar_add_figure}
\alias{ffcar_add_figure}
\title{Create RMarkdown Code for Figures}
\usage{
ffcar_add_figure(chr_index, chr_title = "", chr_caption = "")
}
\arguments{
\item{chr_index}{A character string, assumed to be
an indexed character vector containing the relative
path to the figure (e.g., \code{chr_figure_paths[1]}).}

\item{chr_title}{A character string, an optional
title for the figure.}

\item{chr_caption}{A character vector, an optional
figure caption.}
}
\value{
A character vector.
}
\description{
Function that creates the code to add a figure
(assumed to be a PNG image saved to a local folder)
to a RMarkdown file.
}
