% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Reporting.R
\name{ffcar_add_header}
\alias{ffcar_add_header}
\title{Create RMarkdown Code for Headers}
\usage{
ffcar_add_header(chr_header, int_level = 3)
}
\arguments{
\item{chr_header}{A character string, the header content.}

\item{int_level}{An integer value between 1 and 5, the
level for the header.}
}
\value{
A character vector.
}
\description{
Function that creates a header at a specified
level for a RMarkdown file.
}
\examples{
ffcar_add_header( 'Example' )

}
