% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Reporting.R
\name{ffcar_add_paragraph}
\alias{ffcar_add_paragraph}
\title{Create RMarkdown Code for Paragraph}
\usage{
ffcar_add_paragraph(chr_text)
}
\arguments{
\item{chr_text}{A character vector.}
}
\value{
A character vector.
}
\description{
Function that creates a paragraph for a RMarkdown file.
}
\examples{
ffcar_add_paragraph( 'Hello world' )

}
