% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Reporting.R
\name{ffcar_add_table}
\alias{ffcar_add_table}
\title{Create RMarkdown Code for Tables}
\usage{
ffcar_add_table(chr_index, chr_title, chr_caption, chr_file_name = "")
}
\arguments{
\item{chr_index}{A character string, assumed to be
a pre-existing list containing the specified
table (e.g., \code{lst_example$table}).}

\item{chr_title}{A character string, an optional
title for the table.}

\item{chr_caption}{A character vector, an optional
table caption.}

\item{chr_file_name}{An character string, an optional
file name. If provided, a word document containing the
table will be saved to the 'Downloads' folder.}
}
\value{
A character vector.
}
\description{
Function that creates the code to add a table
(assumed to be a flextable object) to a RMarkdown file.
}
