% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Reporting.R
\name{ffcar_add_table_of_contents}
\alias{ffcar_add_table_of_contents}
\title{Create RMarkdown Code for a Table of Contents}
\usage{
ffcar_add_table_of_contents(chr_headers, chr_sections = "")
}
\arguments{
\item{chr_headers}{A character string, the
text content (i.e., the section header).}

\item{chr_sections}{A character string, the
section number (e.g., \code{'11'}, \code{'22'}).}
}
\value{
A character vector.
}
\description{
Function that creates the code to add
a table of contents linking to the section
headers to a RMarkdown file.
}
