% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Reporting.R
\name{ffcar_conditional_statement}
\alias{ffcar_conditional_statement}
\title{Return Statement Based on Logical Value}
\usage{
ffcar_conditional_statement(lgc_condition, chr_statement = NULL)
}
\arguments{
\item{lgc_condition}{A logical value.}

\item{chr_statement}{A character vector with two strings,
\link{1} the statement to return when \code{FALSE} and \link{2}
the statement to return when \code{TRUE}. Defaults to
a statement on statistical significance.}
}
\value{
A character string.
}
\description{
Function to return one of two possible statements
based on whether a logical value is \code{TRUE} or
\code{FALSE}. Useful, for example, to automatically
generate statements regarding statistical significance.
}
\examples{
ffcar_conditional_statement( FALSE )
ffcar_conditional_statement( TRUE )

# Custom statements
ffcar_conditional_statement(
  FALSE, c( 'non-significant', 'significant' )
)
ffcar_conditional_statement(
  TRUE, c( 'non-significant', 'significant' )
)

}
