% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Reporting.R
\name{ffcar_create_report}
\alias{ffcar_create_report}
\title{Create RMarkdown File for Report}
\usage{
ffcar_create_report(
  chr_file_name,
  chr_date = NULL,
  chr_title = "Analysis Report",
  chr_author = "Kevin Potter"
)
}
\arguments{
\item{chr_file_name}{A character string,
the file name for the report.}

\item{chr_date}{A character vector, the
date for the file name and the date for
the report header, respectively.}

\item{chr_title}{A character string, the
title of the report.}

\item{chr_author}{A character string, the
author(s) of the report.}
}
\value{
An .Rmd file.
}
\description{
Function to initialize a .Rmd file
for a \code{ffcar}-compatible report.
}
