% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R02-Files.R
\name{ffcar_create_report_folder}
\alias{ffcar_create_report_folder}
\title{Create Folders and Subfolders for Report}
\usage{
ffcar_create_report_folder(chr_folder_name)
}
\arguments{
\item{chr_folder_name}{A character string, the
folder name. For compatibility with \code{ffcar}
functions, folder names must start with a letter
and can only contain letters, numbers, underscores,
or periods.}
}
\value{
Creates a folder with relevant subfolders.
}
\description{
Function to create a folder and subfolders for
a standard \code{ffcar}-compatible report.
}
