% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Utilities.R
\name{ffcar_dates}
\alias{ffcar_dates}
\alias{ffcar_ymd_hm}
\alias{ffcar_ymd}
\alias{ffcar_mdy}
\title{Functions to Extract Dates and Times as Strings}
\usage{
ffcar_ymd_hm(chr_sep = "_", chr_split = "-")

ffcar_ymd(chr_sep = "_")

ffcar_mdy(chr_sep = "_")
}
\arguments{
\item{chr_sep}{A character string, the separator
between the year, month, day, hour, and minutes.}

\item{chr_split}{A character string, the separator
between the date and time.}
}
\value{
A character string.
}
\description{
Functions that extract the current date and time
(in UTC) and convert them to formatted strings.
}
\examples{
ffcar_ymd_hm()
ffcar_ymd()
ffcar_mdy()
}
