% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-Tables.R
\name{ffcar_design_table}
\alias{ffcar_design_table}
\title{Design a Table with Summmary Statistics}
\usage{
ffcar_design_table(dtf_data, lst_subsets, ...)
}
\arguments{
\item{dtf_data}{A data frame with values to summarize over.}

\item{lst_subsets}{A list of logical vectors (each matching
in length to the number of rows of \code{dtf_data})
specifying the subsets of data to summarize over. Multiple
logical vectors will result in the creation of separate
columns with summary statistics (one for each logical vector).
The names for the columns of summary statistics will be
set to the list names.}

\item{...}{Additional lists specifying the rows to include
in the table. Each list should be in the format of:
\code{list( 'Label', 'Column name', 'Statistics', Digits,
  'Category' ).} The character string for the row label will
be formatted using the function \link{ffcar_sub_in_content}.
The character string for the column name specifies which column
in \code{dtf_data} summary statistics should be computed
over. The character string for the type of statistics to compute
is passed to the syntax argument of the function
\link{ffcar_summa}. The integer for digits indicates the
number of decimal places summary statistics should be round to.
The final element of the list allows users to specify categories
to match over to create a binary variable for computing proportions,
percentages, and/or frequencies (otherwise the column must already
consist of \code{TRUE} or \code{FALSE} values). Elements of the
list must be specified in order; however, fewer than 5 elements
can be specified. For example, to create a row that consists only
of a label, users can simply provide a list with a single character
string.}
}
\value{
A data frame.
}
\description{
Function to provide a flexible means of creating a table
of summary statistics for a data set (e.g., useful for
creating sample characteristics tables).
}
\examples{
# Example data set
data(mtcars)

# Create summary table
dtf_table <- ffcar_design_table(
  # Data set
  mtcars,
  # Subsets of data (determines # of columns)
  list(
    Overall = rep( TRUE, nrow(mtcars) ),
    Cylinders_4 = mtcars$cyl \%in\% 4,
    Cylinders_6 = mtcars$cyl \%in\% 6,
    Cylinders_8 = mtcars$cyl \%in\% 8
  ),

 # Rows
  # list( Label, Column, Summary, Digits, Category )
  list( 'Sample size',
        'mpg', '[[N]]' ),
  list( 'Miles per gallon; M (SD)',
        'mpg', '[[M]] ([[SD]])', 1 ),
  list( 'Weight (10[[SP3]] lbs); M (SD)',
        'wt', '[[M]] ([[SD]])', 1 ),
  list( 'Transmission; \% automatic (N)',
        'am', '[[P]]\% ([[C]])', 1, 0 )
)
dtf_table[,-(2:5)]

}
