% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R03-Tables.R
\name{ffcar_flextable}
\alias{ffcar_flextable}
\title{Create APA-Style Flextable Object}
\usage{
ffcar_flextable(
  dtf_table,
  chr_header_row_1,
  chr_header_row_2 = NULL,
  chr_header_row_3 = NULL,
  lst_alignment = NULL,
  int_font_size = 11,
  chr_font_family = "Arial",
  lgc_color_rows = NULL,
  chr_row_color = NULL,
  lgc_merge_header_cells = TRUE
)
}
\arguments{
\item{dtf_table}{A data frame with the table content.}

\item{chr_header_row_1}{A character vector matching in
length to the number of columns of \code{dtf_table},
the headers for the columns.}

\item{chr_header_row_2}{A character vector matching in
length to the number of columns of \code{dtf_table},
an optional second line for the column headers.}

\item{chr_header_row_3}{A character vector matching in
length to the number of columns of \code{dtf_table},
an optional third line for the column headers.}

\item{lst_alignment}{A named list consisting of the
elements \code{left}, \code{center}, and/or \code{right},
each consisting of the column indices to apply the
corresponding alignment.}

\item{int_font_size}{An integer value, the font size for
the table.}

\item{chr_font_family}{A character string, the font family.}

\item{lgc_color_rows}{A logical vector matching
in length to the number of rows of \code{dtf_table},
specifying the rows to be colored.}

\item{chr_row_color}{A character string giving the
optional background color for rows. If no input is
provided for \code{lgc_color_rows}, results in
every other row being colored.}

\item{lgc_merge_header_cells}{A logical value; if
\code{TRUE} merges equivalent header cells into a
single cell.}
}
\value{
A \code{flextable} object.
}
\description{
Function that creates an APA-style table
using \code{\link[flextable:flextable]{flextable::flextable()}}.
}
\examples{
data(mtcars)
dtf_example <- mtcars[1:4, 1:3]
ft <- ffcar_flextable(
  dtf_example,
  c( 'Miles per gallon', 'Cylinders', 'Displacement' )
)
ft

}
