% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R02-Files.R
\name{ffcar_list_file_paths}
\alias{ffcar_list_file_paths}
\title{Create List of Files and Folder Paths}
\usage{
ffcar_list_file_paths(chr_folder)
}
\arguments{
\item{chr_folder}{A character string, the path to
the folder with the files, subfolders, and nested
files.}
}
\value{
A list.
}
\description{
Function to create a list detailing all
files, folders, and files within each folder.
Assumes only one level of nested files:
Folder -> Files and subfolders -> Nested files.
}
