% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Reporting.R
\name{ffcar_padded_round}
\alias{ffcar_padded_round}
\title{Round a Value and Pad Number of Trailing Zeros}
\usage{
ffcar_padded_round(num_value, int_digits)
}
\value{
A character string with the rounded values.
}
\description{
Function that rounds a value to a desired number
of digits and
}
\examples{
ffcar_padded_round( c( 1, 1.1, 1.116, 1.1116 ), 2 )

}
