% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R02-Files.R
\name{ffcar_path_to_file}
\alias{ffcar_path_to_file}
\title{Extract Relative Path to File}
\usage{
ffcar_path_to_file(lst_files, chr_partial_name, chr_subfolder = "")
}
\arguments{
\item{lst_files}{A list, output from the
\link{ffcar_list_file_paths} function.}

\item{chr_partial_name}{A character string,
the file name (or part of it) to extract.}

\item{chr_subfolder}{A character string,
the subfolder that contains the file,
if relevant.}
}
\value{
A character string.
}
\description{
Function to extract the relative
path to a file using the output
from the \link{ffcar_list_file_paths}
function and a partial file name.
}
