% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R04-Figures.R
\name{ffcar_png}
\alias{ffcar_png}
\title{Save PNG File}
\usage{
ffcar_png(
  chr_file_name,
  num_width = 5,
  num_height = 5,
  chr_units = "in",
  int_res = 300
)
}
\arguments{
\item{chr_file_name}{A character string, the path for file.}

\item{num_width}{A numeric value, the width of the image
(by default assumed to be in inches).}

\item{num_height}{A numeric value, the height of the image
(by default assumed to be in inches).}

\item{chr_units}{A character string, the units for the
width and height (default is \code{'in' for inches}).}

\item{int_res}{An integer value, the resolution for the image
in ppi (pixels per inch).}
}
\value{
A figure saved to the specified path.
}
\description{
Function to save a PNG file using default options.
}
\details{
Most academic journals expect figures to have dimensions in
either 5 or 7.25 inches and require high-resolution images
(typically 300 ppi).
}
