% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Utilities.R
\name{ffcar_pull_from_matrix}
\alias{ffcar_pull_from_matrix}
\title{Pull Values From Matrix}
\usage{
ffcar_pull_from_matrix(mat_values, int_rows, int_columns, chr_labels = NULL)
}
\arguments{
\item{mat_values}{A matrix.}

\item{int_rows}{An integer vector, the rows to
subset.}

\item{int_columns}{An integer vector, the columns
to subset.}

\item{chr_labels}{A character vector matching in
length to \code{int_columns}, optional new
labels for extracted columns.}
}
\value{
The subset of rows from \code{mat_values}.
}
\description{
Function to extract values over a specified set
of rows and columns from a matrix. Useful, for
example, for pulling results from the
\code{coefficients} matrix returned by
\code{\link[base:summary]{base::summary()}} when working with \code{lm}
or \code{glm} objects.
}
\examples{
# Example data set
data( "mtcars" )
# Multiple regression summary
mat_coefficients <-
  summary( lm( mpg ~ wt + cyl + hp, data = mtcars ) )$coefficients
# Pull estimates, standard errors, and p-values
ffcar_pull_from_matrix(
  mat_coefficients,
  int_rows = 2:4, int_columns = c(1:2, 4),
  chr_labels = c( 'Estimate', 'SE', 'P_value' )
)

# Applied over multiple analyses
lst_summaries <- list(
  wt = summary( lm( mpg ~ wt, data = mtcars ) )$coefficients,
  cyl = summary( lm( mpg ~ cyl, data = mtcars ) )$coefficients,
  hp = summary( lm( mpg ~ hp, data = mtcars ) )$coefficients
)

dtf_summary <- sapply(
  lst_summaries, ffcar_pull_from_matrix,
  int_rows = 2, int_columns = c(1:2, 4),
  chr_labels = c( 'Estimate', 'SE', 'P_value' )
) |> t()


}
