% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Reporting.R
\name{ffcar_unicode_lookup}
\alias{ffcar_unicode_lookup}
\title{Unicode Lookup Table}
\usage{
ffcar_unicode_lookup()
}
\value{
A named character vector with unicode
and associated shorthand labels.
}
\description{
Function that returns useful unicode with
shorthand labels for easy lookup. Note that
for certain systems, the code points
used here may not render properly.
}
\examples{
# Greek letters
ffcar_unicode_lookup() \%with_name\% 'GL'
# Superscripts
ffcar_unicode_lookup() \%with_name\% 'SP'
# Subscripts
ffcar_unicode_lookup() \%with_name\% 'SB'
# Math symbols
ffcar_unicode_lookup() \%with_name\% 'MS'
# Error bar symbols
ffcar_unicode_lookup() \%with_name\% 'EB'

}
