% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Reporting.R
\name{ffcar_update_report_from_list}
\alias{ffcar_update_report_from_list}
\title{Update Report From Content in List}
\usage{
ffcar_update_report_from_list(lst_report, chr_rmd_file, int_max_TOC_level = 2)
}
\arguments{
\item{lst_report}{A list of lists. Each sublist
contains first, a character string specifying
the 'add' function to call, either \code{'header'},
\code{'paragraph'}, \code{'code'}, \code{'figure'},
\code{'table'}, \code{'link'}, or \code{'table_of_contents'}.
Second, the sublist contains a character string
of the form \code{'X_Y'}, where \code{X} is a
label denoting a section and \code{'Y'} is a
number from 1 to 5 denoting the section level.
Finally, the remainder of the list consists
of the inputs to pass to the specified 'add'
function.}

\item{chr_rmd_file}{A character string, the
name for the .Rmd file to update.}

\item{int_max_TOC_level}{A integer value, the maximum
number of tabs to delineate subsections (value between 1 to 6).}
}
\value{
Updated content for the file specified
by \code{chr_rm_file}.
}
\description{
Given a list of lists specifying inputs
for the 'add' functions from \code{Ffcar},
append new content to a \code{.Rmd} file.
}
