% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R05-Reporting.R
\name{ffcar_variables_to_labels}
\alias{ffcar_variables_to_labels}
\title{Convert Variable Names to Labels}
\usage{
ffcar_variables_to_labels(
  chr_variables,
  chr_remove = NULL,
  chr_space = "ZZ",
  int_skip = -1
)
}
\arguments{
\item{chr_variables}{A character vector, the
variable names to convert to nice labels.}

\item{chr_remove}{An optional character vector,
string patterns to remove. Defaults to
\code{'PRD'}, \code{'OUT'}, and \code{'COL'}.}

\item{chr_space}{A character string, the
pattern to replace with an empty space.}

\item{int_skip}{An optional integer vector
with values from 1 to 12 indicating
specifc patterns to skip.}
}
\value{
A character vector of nicely formatted labels.
}
\description{
Function to convert variable names (e.g.,
columns in a data frame, predictors used
in a regression) to nice labels.
}
\examples{
# Parentheses
ffcar_variables_to_labels( 'PRDGroupZZOPPTreatmentCLP' )
# Square brackets
ffcar_variables_to_labels( 'PRDGroupZZOPBTreatmentCLB' )
# Hyphen
ffcar_variables_to_labels( 'PRDShortHPHterm' )
# Colon
ffcar_variables_to_labels( 'PRDGroupCLNZZTreatment' )
# Semicolon
ffcar_variables_to_labels( 'PRDGroupSMCZZTreatment' )
# Pipe symbol
ffcar_variables_to_labels( 'PRDPOPPAPPEBCLP' )
# Plus, minus, and equal signs
ffcar_variables_to_labels( 'PRDXZZEQLZZYZZPLSZZZ' )
# Hash sign
ffcar_variables_to_labels( 'PRDHSHZZofZZcounts' )

}
