% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Utilities.R
\name{\%containing\%}
\alias{\%containing\%}
\title{Operator to Find Element Containing String}
\usage{
lhs \%containing\% rhs
}
\arguments{
\item{lhs, rhs}{Character vectors}
}
\value{
A character vector, the
element(s) in \code{lhs} containing
the \code{rhs} character string.
}
\description{
Operator that returns the element in
a character vector containing
a specified character string.
}
\examples{
s <- c( 'file_1', 'file_2', 'file_3' )
s \%containing\% '2'
# Returns empty set if no matches
s \%containing\% '4'

}
