% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Utilities.R
\name{\%paste\%}
\alias{\%paste\%}
\title{Operator to Combine Two Strings}
\usage{
rhs \%paste\% lhs
}
\arguments{
\item{rhs, lhs}{R objects that can be converted to
character vectors.}
}
\value{
A character vector.
}
\description{
The operator \code{\%paste\%} combines character strings.
}
\details{
The call \code{rhs \%paste\% lhs} is equivalent to
\code{paste0(rhs, lhs)}.
}
\examples{
'Hello' \%paste\% ' ' \%paste\% 'world'

}
