% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Utilities.R
\name{\%pull_sublist\%}
\alias{\%pull_sublist\%}
\title{Operator to Index Sublists in a List}
\usage{
lst_lists \%pull_sublist\% obj_levels
}
\arguments{
\item{obj_levels}{Either a character vector with
the names for the sublists or an integer vector
with the indices.}

\item{lst_content}{A list of lists.}
}
\value{
The element(s) in \code{lhs} whose
names contain the \code{rhs} character string.
}
\description{
Operator that returns the content of a
sublist at a specified level for a list of lists.
}
\examples{
x <- c( A1 = 1, A2 = 2, B1 = 3, B2 = 4 )
x \%with_name\% 'A'
x \%with_name\% 'B'

}
