% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R01-Utilities.R
\name{\%with_name\%}
\alias{\%with_name\%}
\title{Operator to Index Vector by Names}
\usage{
lhs \%with_name\% rhs
}
\arguments{
\item{lhs}{A vector.}

\item{rhs}{A character string.}
}
\value{
The element(s) in \code{lhs} whose
names contain the \code{rhs} character string.
}
\description{
Operator that returns the element(s) in
a named vector whose names partially match
a specified character string.
}
\examples{
x <- c( A1 = 1, A2 = 2, B1 = 3, B2 = 4 )
x \%with_name\% 'A'
x \%with_name\% 'B'

}
